+++ src/main.tsx
@@ -8,8 +8,4 @@ const rootElement = document.querySelector('#root')
 // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 const root = createRoot(rootElement!)
 
-root.render(
-  <StrictMode>
-    <StartScreen />
-  </StrictMode>,
-)
+root.render(<StartScreen />)