+++ src/playlist/PlaylistEditor.tsx
@@ -19,28 +19,27 @@ export default observer(function PlaylistEditor({
       </button>
       {hide ? null : (
         <div className="playlist_editor">
-          <div className="text_header">Enter list of youtube videos</div>
-          <div>
-            <div>
-              <label htmlFor="video">
-                Enter a list of youtube videos separated by newlines, this page
-                will then gather all the videos from the channels that uploaded
-                these videos (I couldn't figure out how to fetch videos from the
-                channel URL itself):
-              </label>
-            </div>
-            <div>
-              <textarea
-                cols={70}
-                rows={5}
-                style={{ maxWidth: '100%' }}
-                id="video"
-                value={model.query}
-                onChange={event => model.setQuery(event.target.value)}
-              />
-            </div>
-            <PlaylistControls model={model} />
+          <div className="text_header">Enter list of:</div>
+          <div className="playlist_margin">
+            <label className="form_description" htmlFor="video">
+              <ul>
+                <li>
+                  URLs for individual youtube videos (URL contains e.g. v=)
+                </li>
+                <li>URLs for youtube playlists (URL contains e.g. list=)</li>
+              </ul>
+              In the first case, the app will download all videos from the
+              channel that uploaded the video, and in the second case, it will
+              download all videos from just that playlist.
+            </label>
+            <textarea
+              id="video"
+              className="form_editor"
+              value={model.query}
+              onChange={event => model.setQuery(event.target.value)}
+            />
           </div>
+          <PlaylistControls model={model} />
         </div>
       )}
     </div>