+++ src/startscreen/ConfirmDialog.tsx
@@ -2,31 +2,36 @@ import { useDialogShown } from '../util'
 
 export default function ConfirmDialog({
   open,
-  setOpen,
+  onClose,
 }: {
   open: boolean
-  setOpen: (arg: boolean) => void
+  onClose: () => void
 }) {
   const ref = useDialogShown(open)
   return (
     <dialog ref={ref} style={{ maxWidth: 500 }}>
       <p>
-        By using this website you agree to usage of the "Privacy Policy" below
+        By using this website you agree to usage of the &quot;Privacy
+        Policy&quot; below
       </p>
       <p>
-        This website (the "app") uses an "API Client" to access the Youtube Data
-        API. The "app" allows third parties to serve content, including
-        advertisements just by virtue of YouTube doing so. See also{' '}
-        <a href="https://www.youtube.com/t/terms">YouTube's Terms of Service</a>{' '}
+        This website (the &quot;app&quot;) uses an &quot;API Client&quot; to
+        access the Youtube Data API. The &quot;app&quot; allows third parties to
+        serve content, including advertisements just by virtue of YouTube doing
+        so. See also{' '}
+        <a href="https://www.youtube.com/t/terms">
+          YouTube&apos;s Terms of Service
+        </a>{' '}
         and{' '}
         <a href="http://www.google.com/policies/privacy">
-          Google's privacy policy
+          Google&apos;s privacy policy
         </a>
         . By using this website you agree to the YouTube Terms of Service. The
-        "app" processes data that the user pastes into the "app", and shares it
-        with the YouTube Data API. User data may also be collected from their
-        device by analytics scripts from the page by the YouTube embedded
-        widget. The "app" itself does not collect any user data or analytics.
+        &quot;app&quot; processes data that the user pastes into the
+        &quot;app&quot;, and shares it with the YouTube Data API. User data may
+        also be collected from their device by analytics scripts from the page
+        by the YouTube embedded widget. The &quot;app&quot; itself does not
+        collect any user data or analytics.
       </p>
       <p>
         I added this consent screen because Google asks users of their API to do
@@ -41,7 +46,7 @@ export default function ConfirmDialog({
       <button
         onClick={() => {
           localStorage.setItem('confirmed', 'true')
-          setOpen(false)
+          onClose()
         }}
       >
         accept