+++ src/startscreen/StartScreen.tsx
@@ -1,27 +1,25 @@
-import { useState } from 'react'
 import App from '../app/App'
-import ConfirmDialog from './ConfirmDialog'
 
 const s = (l: string) => decodeURIComponent(l)
 
 export default function StartScreen() {
-  const [showModal, setShowModal] = useState(!localStorage.getItem('confirmed'))
   const urlParameters = new URLSearchParams(window.location.search)
+
   const ids = s(urlParameters.get('ids') || '')
     .split(',')
     .filter(f => !!f)
-    .map(videoId => `https://www.youtube.com/watch?v=${videoId}`)
-    .join('\n')
+    .map(id => `https://www.youtube.com/watch?v=${id}`)
+  const pids = s(urlParameters.get('pids') || '')
+    .split(',')
+    .filter(f => !!f)
+    .map(id => `https://www.youtube.com/watch?list=${id}`)
+
   const playlist = s(urlParameters.get('playlist') || 'default')
 
   return (
-    <>
-      <ConfirmDialog open={showModal} setOpen={setShowModal} />
-      <App
-        initialQuery={ids}
-        initialPlaylist={playlist}
-        showPrivacyPolicy={() => setShowModal(true)}
-      />
-    </>
+    <App
+      initialQuery={[...ids, ...pids].join('\n')}
+      initialPlaylist={playlist}
+    />
   )
 }